feature_store/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pandas==1.3.3
numpy==1.21.2
python-dotenv==1.0.1
